reduce code duplication

+++ include/jdx/util/PeakTableParser.hpp
@@ -13,7 +13,7 @@ namespace sciformats::jdx::util
 class PeakTableParser
 {
 public:
-    explicit PeakTableParser(std::istream& iStream, unsigned int numVariables);
+    explicit PeakTableParser(std::istream& iStream, size_t numVariables);
     /**
      * @brief Next table item.
      * @note Assumes that a peak tuple does not span multiple lines, but one
@@ -26,7 +26,7 @@ public:
 
 private:
     std::istream& m_istream;
-    unsigned int m_numVariables;
+    size_t m_numVariables;
     bool m_isPastInitialComment;
     std::string m_currentLine;
     size_t m_currentPos;
